close #125 cambio de nombre carpeta class y añadida función para recuperar un evento y mostrarlo con la nueva plantilla
