+++ _dt/css/main.css
@@ -15,7 +15,7 @@ GNU General Public License for more details.
 */
 
 body {
-    padding-top: 80px;
+    padding-top: 70px;
     font-family: verdana;
     font-size: 12px;
     background-color: #eee;
@@ -65,12 +65,6 @@ html {
     min-height: 100%;
 }
 
-body {
-    /* Margin bottom by footer height */
-    
-    margin-bottom: 60px;
-}
-
 .footer {
     position: absolute;
     bottom: 0;
@@ -85,13 +79,15 @@ body {
 /* Not required for template or sticky footer method. */
 
 body > .container {
-    padding: 60px 15px 0;
+    /*padding: 60px 15px 0;*/
+  padding: 0;
+  margin: 0;
 }
 
 .container .text-muted {
     margin: 20px 0;
 }
-
+ 
 .footer > .container {
     padding-right: 15px;
     padding-left: 15px;
@@ -100,6 +96,7 @@ body > .container {
 code {
     font-size: 80%;
 }
+
 #cabecera-tabla-ponentes {
     background-color:#C11F23;
     color:white;
@@ -107,4 +104,61 @@ code {
 
 #body-tabla-ponentes {
     background-color: #fff;
+}
+
+/*************************************************/
+/*********** templateEvents estilos **************/
+/*************************************************/
+.menu-content {
+  position: fixed;
+  top: 60px;
+  bottom: 0;
+  left: 0;
+  z-index: 1000;
+  display: block;
+  padding: 20px;
+  overflow-x: hidden;
+  overflow-y: auto; 
+  background-color: white;
+}
+
+.menu-content .menu .menu-nav {
+  margin-right: -21px; 
+  margin-bottom: 20px;
+  margin-left: -20px;
+}
+
+.menu-content .menu .menu-nav li > a {
+  background: none;
+  padding: 5px 0 5px 15px;
+  font-size: 14px;
+  
+}
+
+.menu-content .menu .menu-nav .titulo-menu > a {
+  color: #C11F23;
+  font-weight: bold;
+}
+
+.menu-content .menu .menu-nav .option-menu > a {
+  color: #959598;
+}
+
+.menu-content .menu .menu-nav .option-menu > a:hover,
+.menu-content .menu .menu-nav .option-menu > a:focus {
+  color: #3b3b3c;
+}
+
+.menu-content .menu .menu-nav .active > a {
+  background: #f5f5f5;
+}
+
+.event-content {
+  padding: 25px;
+}
+
+.event-content .titular-h1 {
+  margin: 0;
+  padding: 0;
+  font-size: 30px;
 }
