+++ _oo/classes/asistente.py
