+++ templates/index.html
@@ -69,15 +69,14 @@ GNU General Public License for more details.
     </nav>
     <!--/navegador -->{% endblock %}
 
-     Begin page content
     <div class="container">
         {% block content %} {% endblock %}
     </div>
         {% block footer %}
-    <footer class="footer">
+    <!--<footer class="footer">
         <div class="container col-xs-6">
             <h3>Sobre Evenge</h3>
-            <p>Pin a fixed-height footer to the bottom of the viewport in desktop browsers with this custom HTML and CSS.Pin a fixed-height footer to the bottom of the viewport in desktop browsers with this custom HTML and CSS.Pin a fixed-height footer to the bottom of the viewport in desktop browsers with this custom HTML and CSS.Pin a fixed-height footer to the bottom of the viewport in desktop browsers with this custom HTML and CSS.</p>
+            <p></p>
         </div>
         <div class="container col-xs-6">
             <h3>Contacto</h3>
@@ -94,7 +93,7 @@ GNU General Public License for more details.
                     a</li>
             </ul>
         </div>
-    </footer>
+    </footer>-->
         {% endblock %}
 
             {% block scriptFooter %}
